requirements.txt
streamlit
opencv-python-headless
scikit-learn
numpy
Pillow
